requirements.txt
fastapi
uvicorn
twilio
elevenlabs
python-dotenv
openai
python-multipart
requests

